Refactor Web3Manager to typescript
