+++ libs/package.json
@@ -72,6 +72,7 @@
     "@tsconfig/node16-strictest": "1.0.0",
     "@types/async-retry": "1.4.3",
     "@types/bs58": "4.0.1",
+    "@types/eth-sig-util": "^2.1.1",
     "@types/expect": "24.3.0",
     "@types/hashids": "2.0.1",
     "@types/mocha": "9.1.0",
@@ -87,6 +88,7 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "5.2.0",
+    "ethereumjs-wallet": "^1.0.2",
     "mocha": "9.2.2",
     "nock": "13.1.2",
     "nyc": "15.1.0",
