+++ libs/src/services/ethWeb3Manager/EthWeb3Manager.ts
@@ -1,26 +1,18 @@
 import Web3 from '../../web3'
 import type Web3Type from 'web3'
-import {
-  MultiProvider,
-  estimateGas,
-  Providers,
-  ContractMethod
-} from '../../utils'
+import { MultiProvider, estimateGas, ContractMethod } from '../../utils'
 import { Transaction as EthereumTx } from 'ethereumjs-tx'
 import retry from 'async-retry'
 import type { IdentityService, RelayTransaction } from '../identity'
 import type { Hedgehog } from '@audius/hedgehog'
 import type { AxiosError } from 'axios'
+import type { Web3Config } from '../web3Manager'
+import type Wallet from 'ethereumjs-wallet'
 
 const MIN_GAS_PRICE = Math.pow(10, 9) // 1 GWei, ETH minimum allowed gas price
 const HIGH_GAS_PRICE = 250 * MIN_GAS_PRICE // 250 GWei
 const DEFAULT_GAS_PRICE = 100 * MIN_GAS_PRICE // 100 Gwei is a reasonably average gas price
 const MAX_GAS_LIMIT = 5000000 // We've seen prod tx's take up to 4M. Set to the highest we've observed + a buffer
-
-type Web3Config = {
-  ownerWallet: string
-  providers: Providers
-}
 
 /** Singleton state-manager for Audius Eth Contracts */
 export class EthWeb3Manager {
@@ -28,7 +20,7 @@ export class EthWeb3Manager {
   web3: Web3Type
   identityService: IdentityService
   hedgehog: Hedgehog
-  ownerWallet: string | undefined
+  ownerWallet: Wallet | undefined
 
   constructor(
     web3Config: Web3Config,
@@ -63,6 +55,7 @@ export class EthWeb3Manager {
 
   getWalletAddress() {
     if (this.ownerWallet) {
+      // @ts-expect-error TODO extend ethereum-js-wallet to include toLowerCase
       return this.ownerWallet.toLowerCase()
     }
     throw new Error('Owner wallet not set')
@@ -91,7 +84,7 @@ export class EthWeb3Manager {
       txGasLimit ??
       (await estimateGas({
         method: contractMethod,
-        from: this.ownerWallet as string,
+        from: this.ownerWallet,
         gasLimitMaximum: MAX_GAS_LIMIT
       }))
     if (contractAddress && privateKey) {
@@ -148,7 +141,7 @@ export class EthWeb3Manager {
 
     const gasPrice = parseInt(await this.web3.eth.getGasPrice())
     return contractMethod.send({
-      from: this.ownerWallet as string,
+      from: this.ownerWallet,
       gas: gasLimit,
       gasPrice: gasPrice
     })
@@ -170,7 +163,7 @@ export class EthWeb3Manager {
     contractMethod: ContractMethod,
     contractAddress: string,
     ownerWallet: string,
-    relayerWallet: string,
+    relayerWallet: Wallet,
     txRetries = 5,
     txGasLimit: number | null = null
   ): Promise<RelayTransaction['resp'] | undefined> {
@@ -227,7 +220,7 @@ export class EthWeb3Manager {
   async getRelayMethodParams(
     contractAddress: string,
     contractMethod: ContractMethod,
-    relayerWallet: string
+    relayerWallet: Wallet
   ) {
     const encodedABI = contractMethod.encodeABI()
     const gasLimit = await estimateGas({
