+++ libs/src/services/web3Manager/XMLHttpRequest.ts
@@ -1,0 +1,9 @@
+let XMLHttpRequestRef: typeof window.XMLHttpRequest
+
+if (typeof window === 'undefined' || window === null) {
+  XMLHttpRequestRef = require('xmlhttprequest').XMLHttpRequest
+} else {
+  XMLHttpRequestRef = window.XMLHttpRequest
+}
+
+export { XMLHttpRequestRef as XMLHttpRequest }
