+++ libs/src/services/web3Manager/index.ts
@@ -1,0 +1,2 @@
+export * from './Web3Config'
+export * from './Web3Manager'
